requirements.txt

django~=3.0.7

djangorestframework==3.11.0
django-restful-admin==1.1.0
django-split-settings==1.0.1
django-cors-headers==3.4.0
# mysqlclient==1.3.12
djangorestframework-jwt==1.11.0
# django-jet==1.0.8
bcrypt
requests
django-markdownx==3.0.1


# amqp==2.6.0
# certifi==2020.6.20
# chardet==3.0.4
# django-nameko-standalone==1.3.2
# dnspython==1.16.0
# eventlet==0.25.2
# greenlet==0.4.16
# idna==2.9
# importlib-metadata==1.6.1
# kombu==4.6.11
# mock==4.0.2
# monotonic==1.5
# more-itertools==8.4.0
# nameko==2.12.0
# path.py==12.4.0
# pytz==2020.1
# PyYAML==5.3.1
# requests==2.24.0
# six==1.15.0
# sqlparse==0.3.1
# urllib3==1.25.9
# vine==1.3.0
# Werkzeug==1.0.1
# wrapt==1.11.2
# zipp==3.1.0